requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic-settings==2.2.1
email-validator==2.1.1
apscheduler==3.10.4
